Move constants to index file
