+++ projects/js-packages/publicize-components/src/components/social-previews/linkedin.js
@@ -3,8 +3,7 @@ import { useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import React from 'react';
 import useSocialMediaMessage from '../../hooks/use-social-media-message';
-import { SOCIAL_STORE_ID } from '../../social-store';
-import { CONNECTION_SERVICE_LINKEDIN } from '../../social-store/constants';
+import { SOCIAL_STORE_ID, CONNECTION_SERVICE_LINKEDIN } from '../../social-store';
 
 /**
  * The linkedin tab component.
