--- projects/js-packages/publicize-components/src/social-store/constants.js
@@ -1,7 +1,0 @@
-export const CONNECTION_SERVICE_FACEBOOK = 'facebook';
-export const CONNECTION_SERVICE_INSTAGRAM_BUSINESS = 'instagram-business';
-export const CONNECTION_SERVICE_LINKEDIN = 'linkedin';
-export const CONNECTION_SERVICE_MASTODON = 'mastodon';
-export const CONNECTION_SERVICE_NEXTDOOR = 'nextdoor';
-export const CONNECTION_SERVICE_TUMBLR = 'tumblr';
-export const CONNECTION_SERVICE_TWITTER = 'twitter';
