+++ projects/js-packages/publicize-components/src/social-store/index.js
@@ -19,5 +19,13 @@ export const SOCIAL_STORE_CONFIG = {
 		{},
 };
 
+export const CONNECTION_SERVICE_FACEBOOK = 'facebook';
+export const CONNECTION_SERVICE_INSTAGRAM_BUSINESS = 'instagram-business';
+export const CONNECTION_SERVICE_LINKEDIN = 'linkedin';
+export const CONNECTION_SERVICE_MASTODON = 'mastodon';
+export const CONNECTION_SERVICE_NEXTDOOR = 'nextdoor';
+export const CONNECTION_SERVICE_TUMBLR = 'tumblr';
+export const CONNECTION_SERVICE_TWITTER = 'twitter';
+
 const store = createReduxStore( SOCIAL_STORE_ID, SOCIAL_STORE_CONFIG );
 register( store );
